break some functions out
